miror changes and proper roll based access for all data

+++ shyam_sunder_reddy/ust_employee_task_management/backend/routers/employee_api.py
@@ -9,6 +9,9 @@ employee_router = APIRouter(prefix="/Employee", tags=["Employee"])
 @employee_router.get("/getall", response_model=List[EmployeeReqRes])
 def get_all(role: str, user=Depends(get_current_user)):
     try:
+        if role not in user.role:
+            raise HTTPException(status_code=400,detail="you dont have the access of mentioned role")
+        
         employees = get_all_employees(role, user)
         if not employees:
             raise HTTPException(status_code=404, detail="No employees found")
@@ -23,6 +26,8 @@ def add_new_employee(role: str, new_emp: EmployeeReqRes, user=Depends(get_curren
     try:
         if role != "Admin":
             raise HTTPException(status_code=403, detail="Only Admin can create employees.")
+        if role not in user.role:
+            raise HTTPException(status_code=400,detail="you dont have the access of mentioned role")
         
         new_employee = add_employee(new_emp, role, user)
         return {"detail": "Employee Added Successfully", "employee": new_employee}
@@ -34,6 +39,9 @@ def add_new_employee(role: str, new_emp: EmployeeReqRes, user=Depends(get_curren
 @employee_router.get("/get", response_model=EmployeeReqRes)
 def get_by_id(id: int, role: str, user=Depends(get_current_user)):
     try:
+        if role not in user.role:
+            raise HTTPException(status_code=400,detail="you dont have the access of mentioned role")
+        
         emp = get_by_employee_id(id, role, user)
         return emp
     except HTTPException as e:
@@ -42,8 +50,11 @@ def get_by_id(id: int, role: str, user=Depends(get_current_user)):
         raise HTTPException(status_code=500, detail=f"Internal Server Error: {str(e)}")
 
 @employee_router.put("/update")
-def update_employee_data(id: int, new_data: dict, role: str, user=Depends(get_current_user)):
+def update_employee_data(id: int, new_data: EmployeeReqRes, role: str, user=Depends(get_current_user)):
     try:
+        if role not in user.role:
+            raise HTTPException(status_code=400,detail="you dont have the access of mentioned role")
+        
         updated_emp = update_employee(id, new_data, role, user)
         return {"detail": "Employee Updated Successfully", "employee": updated_emp}
     except HTTPException as e:
@@ -54,6 +65,9 @@ def update_employee_data(id: int, new_data: dict, role: str, user=Depends(get_cu
 @employee_router.delete("/delete")
 def delete_employee_by_id(id: int, role: str, user=Depends(get_current_user)):
     try:
+        if role not in user.role:
+            raise HTTPException(status_code=400,detail="you dont have the access of mentioned role")
+        
         delete_response = delete_employee(id, role, user)
         return delete_response
     except HTTPException as e:
